@@ -45,20 +45,20 @@
         "tslib": "^2.4.0"
     },
     "devDependencies": {
-        "@rollup/plugin-commonjs": "^22.0.0",
-        "@rollup/plugin-node-resolve": "^13.2.1",
-        "@rollup/plugin-typescript": "^8.3.2",
+        "@rollup/plugin-commonjs": "^25.0.1",
+        "@rollup/plugin-node-resolve": "^15.1.0",
+        "@rollup/plugin-terser": "^0.4.3",
+        "@rollup/plugin-typescript": "^11.1.1",
         "@types/d3-axis": "^3.0.1",
         "@types/d3-color": "^3.0.2",
         "@types/d3-scale": "^4.0.1",
         "@types/d3-selection": "^3.0.1",
-        "@types/jest": "^27.4.1",
+        "@types/jest": "^29.5.2",
         "http-server": "^14.0.0",
-        "jest": "^27.5.1",
-        "rollup": "^2.71.1",
-        "rollup-plugin-terser": "^7.0.2",
-        "ts-jest": "^27.1.4",
-        "tsd": "^0.20.0",
-        "typescript": "^4.6.4"
+        "jest": "^29.5.0",
+        "rollup": "^3.25.0",
+        "ts-jest": "^29.1.0",
+        "tsd": "^0.28.1",
+        "typescript": "^5.1.3"
     }
 }
