@@ -1,9 +1,9 @@
 import resolve from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
 import typescript from '@rollup/plugin-typescript'
-import { terser } from "rollup-plugin-terser";
+import terser from "@rollup/plugin-terser";
 
-const pkg = require('./package.json')
+import pkg from './package.json' assert {type: 'json'};
 
 const config = {
     input: `src/index.ts`,
